requirements.txt
certifi==2023.7.22
charset-normalizer==3.1.0
idna==3.4
orjson==3.8.12
psycopg2-binary==2.9.6
requests==2.31.0
urllib3==2.0.7
